split up remaining tests and update imports
