+++ packages/calcite-components/src/tests/commonTests/floatingUI.ts
@@ -1,4 +1,3 @@
-/* eslint-disable jest/no-conditional-expect -- Using conditional logic in a confined test helper to handle specific scenarios, reducing duplication, balancing test readability and maintainability. **/
 /* eslint-disable jest/no-export -- Util functions are now imported to be used as `it` blocks within `describe` instead of assertions within `it` blocks. */
 import { toHaveNoViolations } from "jest-axe";
 import type { JSX } from "../../components";
