+++ packages/calcite-components/src/tests/commonTests/labelable.ts
@@ -1,4 +1,3 @@
-/* eslint-disable jest/no-conditional-expect -- Using conditional logic in a confined test helper to handle specific scenarios, reducing duplication, balancing test readability and maintainability. **/
 /* eslint-disable jest/no-export -- Util functions are now imported to be used as `it` blocks within `describe` instead of assertions within `it` blocks. */
 import { E2EPage, newE2EPage } from "@stencil/core/testing";
 import { toHaveNoViolations } from "jest-axe";
@@ -10,8 +9,8 @@ import {
   isHTML,
   getTag,
   getTagOrHTMLWithBeforeContent,
-} from "./../commonTestUtils/setupForTests";
-import { FocusableOptions } from "./../commonTestUtils/focusable";
+} from "./../commonTests/setupForTests";
+import { FocusableOptions } from "./focusable";
 
 expect.extend(toHaveNoViolations);
