+++ packages/calcite-components/src/tests/commonTests/setupForTests.ts
@@ -1,4 +1,3 @@
-/* eslint-disable jest/no-conditional-expect -- Using conditional logic in a confined test helper to handle specific scenarios, reducing duplication, balancing test readability and maintainability. **/
 /* eslint-disable jest/no-export -- Util functions are now imported to be used as `it` blocks within `describe` instead of assertions within `it` blocks. */
 import { E2EPage, newE2EPage } from "@stencil/core/testing";
 import { toHaveNoViolations } from "jest-axe";
@@ -91,4 +90,8 @@ export function getTagOrHTMLWithBeforeContent(componentTestSetup: TagOrHTML | Ta
     tagOrHTML: componentTestSetup.tagOrHTML,
     beforeContent: componentTestSetup.beforeContent,
   };
+}
+
+export function propToAttr(name: string): string {
+  return name.replace(/([a-z0-9])([A-Z])/g, "$1-$2").toLowerCase();
 }
